requirements.txt
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-betterforms==1.2
django-cors-headers==3.0.1
django-emoji-picker==0.0.4
django-heroku==0.3.1
django-pagedown==1.0.6
django-upload-validator==1.0.2
djangorestframework==3.9.4
djongo==1.2.32
Faker==1.0.5
future==0.17.1
gunicorn==19.9.0
idna==2.8
Markdown==3.1
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyJWT==1.7.1
pymongo==3.8.0
python-dateutil==2.8.0
python-magic==0.4.15
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
serializers==0.2.4
six==1.12.0
social-auth-core==3.1.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.24.3
whitenoise==4.1.2
